+++ form/product.go
@@ -3,17 +3,17 @@ package form
 import "time"
 
 type ProductListQueryForm struct {
-	ProjectID *int `form:"projectID" binding:"required"`
+	ProjectID *int64 `form:"projectID" binding:"required"`
 }
 
 type Product struct {
-	ID                  int       `json:"id"`
+	ID                  int64     `json:"id"`
 	Name                string    `json:"name"`
 	Status              int       `json:"status"`
 	SupplierName        string    `json:"supplierName,omitempty"`
 	SupplierPrice       int       `json:"supplierPrice,omitempty"`
 	SupplierProductName string    `json:"supplierProductName,omitempty"`
-	SupplierProductID   int       `json:"supplierProductID,omitempty"` //如果没有，不序列化
+	SupplierProductID   int64     `json:"supplierProductID,omitempty"` //如果没有，不序列化
 	FacePrice           int       `json:"facePrice"`
 	Price               int       `json:"price"`
 	Model               int       `json:"model"`
@@ -27,11 +27,11 @@ type ProductListQueryResp struct {
 }
 
 type ProductSupplier struct {
-	ID                  int    `json:"id"`
-	SupplierID          int    `json:"supplierID"`
+	ID                  int64  `json:"id"`
+	SupplierID          int64  `json:"supplierID"`
 	SupplierName        string `json:"supplierName"`
 	SupplierCode        string `json:"supplierCode"`
-	SupplierProductID   int    `json:"supplierProductID"`
+	SupplierProductID   int64  `json:"supplierProductID"`
 	SupplierProductName string `json:"supplierProductName"`
 	SupplierProductCode string `json:"supplierProductCode"`
 	FacePrice           int    `json:"facePrice"`
